moving initial code to src
